pyproject.toml
[tool.poetry]
name = "multimedeval"
version = "1.0.0"
description = "A Python tool to evaluate the performance of VLM on the medical domain."
authors = ["Corentin Royer"]
readme = "README.md"
homepage = "https://github.com/corentin-ryr/MultiMedEval"
repository = "https://github.com/corentin-ryr/MultiMedEval"
documentation = "https://github.com/corentin-ryr/MultiMedEval"
keywords = ["evaluation", "medical", "vlm"]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
packages = [{include="multimedeval"}]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
torch = "*"
transformers = "*"
datasets = ">=2.16"
nltk = "*"
torchmetrics = "*"
gdown = "*"
scikit-learn = "1.3.2" # For pickling warning messages
statsmodels = "*"
bert_score = "*"
pydicom = "*"
kaggle = "*"
medmnist = "*"
Pillow = ">=10.2.0"
types-requests = "*"
protobuf = "*"
nibabel = "*"
spacy = ">=3.6.0,<3.8.4"
jsonpickle = "*"
h5py = "*"
dotmap = "^1.3.30"
openai = "^1.93.2"
tqdm = "^4.67.1"
pydantic-ai = "^0.4.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
pytest-order = "^1.0"
plotly = "^5.0"
tabulate = ">=0.9.0"
types-tabulate = "^0.8"
colour = "^0.1"
flake8 = "^6.0"
black = "^23.0"
pylint = "^2.0"
pre-commit = "^3.0"
mypy = "^1.0"
isort = "^5.0"
flake8-bugbear = "^23.0"
flake8-comprehensions = "^3.0"
flake8-executable = "^2.0"
flake8-pyi = "*"
appdirs = "^1.4.4"
kaleido = "0.2.1"

[tool.poetry.group.ctclip.dependencies]
einops = "*"
vector-quantize-pytorch = "^1.20.9"



[tool.pytest.ini_options]
addopts = "--cov=multimedeval"
testpaths = "tests"

[tool.coverage.run]
omit = ["multimedeval/chexbert/*", "multimedeval/tqdm_loggable.py", "multimedeval/radgraph/*"]

[tool.isort]
profile = "black"
line_length = 88
py_version=38

[tool.mypy]
pretty = true
# strict = true
ignore_missing_imports = true
no_implicit_optional = true
python_version = "3.8"

[[tool.mypy.overrides]]
module = "numpy"
ignore_missing_imports = true